Change types and fix hooks

+++ Project IC/Project IC/Utils/VMT/VMT.h
@@ -3,6 +3,9 @@
 //std::unique_ptr
 #include <memory>
 
+//std::memcpy
+#include <cstring>
+
 //Written by a pleb
 namespace VMT
 {
@@ -32,7 +35,7 @@ namespace VMT
 
 		template <typename FN>
 		inline FN Original(const int index)	{ return reinterpret_cast<FN>(original[index]); }
-		inline void Hook(const int index, void *function)	{ current[index] = reinterpret_cast<unsigned int>(function); }
+		inline void Hook(const int index, void *function)	{ current[index] = reinterpret_cast<uintptr_t>(function); }
 		inline void Unhook(const int index)					{ current[index] = original[index]; }
 		inline bool Initialized()                           { return size && original != nullptr; }
 		inline void RestoreTable()                          { if (Initialized()) *baseclass = original; }
